refactor: remove old rule simulation service and controller, integrate excluded types into simulation service
