--- backend/src/services/rule-simulation/dto/rule-simulation.dto.ts
@@ -1,19 +1,0 @@
-import { IsString, IsNotEmpty, IsOptional } from 'class-validator';
-
-export class ExcludedTypeProps {
-    @IsOptional()
-    @IsString()
-    masking_id?: null | string;
-
-    @IsString()
-    @IsNotEmpty()
-    txtp?: string;
-
-    @IsString()
-    @IsNotEmpty()
-    txtp_version?: string;
-
-    @IsString()
-    @IsNotEmpty()
-    record_status?: string;
-}
