+++ backend/src/services/simulation/simulation.service.ts
@@ -1,13 +1,13 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { AdminServiceClient } from '../admin-service-client';
 import type { AuthenticatedUser } from '../auth/auth.types';
-import type { SimulationListResponseDto, CreateSimulationDto, CreateSimulationResponseDto } from './dto/simulation.dto';
+import type { SimulationListResponseDto, CreateSimulationDto, CreateSimulationResponseDto, ExcludedTypeProps } from './dto/simulation.dto';
 
 @Injectable()
 export class SimulationService {
   private readonly logger = new Logger(SimulationService.name);
 
-  constructor(private readonly adminServiceClient: AdminServiceClient) {}
+  constructor(private readonly adminServiceClient: AdminServiceClient) { }
 
   async getAllSimulations(
     offset: number,
@@ -30,6 +30,17 @@ export class SimulationService {
       return await this.adminServiceClient.createSimulation(body, user.token.tokenString);
     } catch (error) {
       this.logger.error(`Error creating simulation: ${error instanceof Error ? error.message : String(error)}`);
+      throw error;
+    }
+  }
+
+  async excludedTypes(token: string): Promise<ExcludedTypeProps[]> {
+    try {
+      const response = await this.adminServiceClient.getExcludedTypes(token);
+      return response
+    } catch (error) {
+      const err = error as Error;
+      this.logger.error(`Error fetching excluded types: ${err.message}`);
       throw error;
     }
   }
